llm-api-wrapper/requirements.txt
pydantic>=2.0.0
httpx>=0.24.0
openai>=1.0.0
anthropic>=0.3.0
google-genai>=0.1.0
python-dotenv>=1.0.0
streamlit>=1.24.0
rich>=13.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0